Simplify conversion test using hypothesis
