+++ tests/conftest.py
@@ -1,9 +1,14 @@
 """Test confiugration file."""
 
 import pytest
+from hypothesis import strategies
 
+from hdate.hebrew_date import HebrewDate, Months
 from hdate.holidays import HolidayDatabase
 from hdate.location import Location
+
+MIN_YEAR = 4000
+MAX_YEAR = 6000
 
 
 @pytest.fixture(scope="session")
@@ -47,3 +52,14 @@ def location(request: pytest.FixtureRequest) -> Location:
     if request.param not in locations:
         raise ValueError(f"Invalid location: {request.param}")
     return locations[request.param]
+
+
+@strategies.composite
+def valid_hebrew_date(draw: strategies.DrawFn) -> HebrewDate:
+    """Generate a valid Hebrew date."""
+    year = draw(strategies.integers(min_value=MIN_YEAR, max_value=MAX_YEAR))
+    month = draw(strategies.sampled_from(Months.in_year(year)))
+    days = HebrewDate(year).days_in_month(month)
+    day = draw(strategies.integers(min_value=1, max_value=days))
+
+    return HebrewDate(year, month, day)
